Moving FindCasterSpell into CraftingLogic
